+++ contracts/policies/CombinedPoliciesPolicy.sol
@@ -25,9 +25,9 @@ contract CombinedPoliciesPolicy is FirewallPolicyBase {
         authorizedExecutors[_firewallAddress] = true;
     }
 
-    function preExecution(address consumer, address sender, bytes calldata data, uint value) external isAuthorized(consumer) {
+    function preExecution(address consumer, address sender, bytes calldata data, uint256 value) external isAuthorized(consumer) {
         bool[] memory currentResult = new bool[](policies.length);
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             IFirewallPolicy policy = IFirewallPolicy(policies[i]);
             try policy.preExecution(consumer, sender, data, value) {
                 currentResult[i] = true;
@@ -38,10 +38,10 @@ contract CombinedPoliciesPolicy is FirewallPolicyBase {
         currentResults.push(currentResult);
     }
 
-    function postExecution(address consumer, address sender, bytes calldata data, uint value) external isAuthorized(consumer) {
+    function postExecution(address consumer, address sender, bytes calldata data, uint256 value) external isAuthorized(consumer) {
         bool[] memory currentResult = currentResults[currentResults.length - 1];
         currentResults.pop();
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             IFirewallPolicy policy = IFirewallPolicy(policies[i]);
             try policy.postExecution(consumer, sender, data, value) {
                 // Do nothing
@@ -55,12 +55,12 @@ contract CombinedPoliciesPolicy is FirewallPolicyBase {
 
     function setAllowedCombinations(address[] calldata _policies, bool[][] calldata _allowedCombinations) external onlyRole(POLICY_ADMIN_ROLE) {
         // Reset all combinations to false
-        for (uint i = 0; i < allowedCombinationHashes.length; i++) {
+        for (uint256 i = 0; i < allowedCombinationHashes.length; i++) {
             isAllowedCombination[allowedCombinationHashes[i]] = false;
         }
         allowedCombinationHashes = new bytes32[](_allowedCombinations.length);
         // Set all new combinations to true
-        for (uint i = 0; i < _allowedCombinations.length; i++) {
+        for (uint256 i = 0; i < _allowedCombinations.length; i++) {
             isAllowedCombination[keccak256(abi.encodePacked(_allowedCombinations[i]))] = true;
             allowedCombinationHashes[i] = (keccak256(abi.encodePacked(_allowedCombinations[i])));
         }
