+++ contracts/proxy/FirewallProxyIntercept.sol
@@ -71,7 +71,7 @@ contract FirewallProxyIntercept is TransparentUpgradeableProxy, IFirewallConsume
             return;
         }
 
-        uint value;
+        uint256 value;
         // We do this because msg.value can only be accessed in payable functions.
         assembly {
             value := callvalue()
